begin work on subsample loop lengths

+++ src/looprecording/looplayertracker.cpp
@@ -48,7 +48,7 @@ LoopLayerTracker::~LoopLayerTracker() {
     delete m_pLoopDeck1Play;
 }
 
-void LoopLayerTracker::addLoopLayer(QString path, unsigned int length) {
+void LoopLayerTracker::addLoopLayer(QString path, double length) {
     LayerInfo* layer = new LayerInfo();
     layer->path = path;
     layer->length = length;
@@ -126,7 +126,7 @@ QString LoopLayerTracker::getCurrentPath() {
     }
 }
 
-void LoopLayerTracker::setCurrentLength(unsigned int length) {
+void LoopLayerTracker::setCurrentLength(double length) {
 
 }
